Added functionality to save notes manually and/or automatically

+++ src/server/api/routers/notes.ts
@@ -1,25 +1,71 @@
 import { z } from "zod";
 import { createTRPCRouter, protectedProcedure } from "../trpc";
+import DMP from "diff-match-patch";
+import { TRPCError } from "@trpc/server";
+import { Prisma, PrismaClient } from "@prisma/client";
+import { Session } from "next-auth";
+
+const diffSchema = z.tuple([z.number(), z.string()]);
+
+const patchesSchema = z.array(
+  z.object({
+    diffs: diffSchema.array(),
+    start1: z.number().nullable(),
+    start2: z.number().nullable(),
+    length1: z.number(),
+    length2: z.number(),
+  })
+);
+
+// get current note content, check ownership
+async function getFullNote(
+  prisma: PrismaClient,
+  session: Session,
+  noteId: string
+) {
+  const note = await prisma.note.findFirst({
+    where: { id: noteId, userId: session.user.id },
+    include: { tags: true },
+  });
+  if (!note) return null;
+  return note;
+}
 
 export const notesRouter = createTRPCRouter({
   getSingle: protectedProcedure
     .input(z.object({ id: z.string().cuid() }))
-    .query(async ({ ctx, input }) => {
-      const note = await ctx.prisma.note.findUnique({
-        where: { id: input.id },
-        include: {
+    .query(({ ctx, input }) => {
+      return ctx.prisma.note.findFirst({
+        where: { id: input.id, userId: ctx.session.user.id },
+        select: {
+          id: true,
+          createdAt: true,
+          lastUpdated: true,
+          title: true,
+          userId: true,
           tags: { orderBy: { createdAt: "asc" } },
         },
       });
-      if (!note) return null;
-      if (note.userId !== ctx.session.user.id) return null;
-      return note;
+    }),
+
+  getSingleContent: protectedProcedure
+    .input(z.object({ id: z.string().cuid() }))
+    .query(async ({ ctx, input }) => {
+      return ctx.prisma.note.findFirst({
+        where: { id: input.id, userId: ctx.session.user.id },
+        select: { content: true, userId: true },
+      });
     }),
 
   getAll: protectedProcedure.query(({ ctx }) => {
     return ctx.prisma.note.findMany({
       where: { userId: ctx.session.user.id },
-      include: {
+      select: {
+        id: true,
+        createdAt: true,
+        lastUpdated: true,
+        title: true,
+        userId: true,
         tags: { orderBy: { createdAt: "asc" } },
       },
     });
@@ -36,7 +82,12 @@ export const notesRouter = createTRPCRouter({
           },
           userId: ctx.session.user.id,
         },
-        include: {
+        select: {
+          id: true,
+          createdAt: true,
+          lastUpdated: true,
+          title: true,
+          userId: true,
           tags: { orderBy: { createdAt: "asc" } },
         },
       });
@@ -45,9 +96,20 @@ export const notesRouter = createTRPCRouter({
   delete: protectedProcedure
     .input(z.object({ id: z.string().cuid() }))
     .mutation(async ({ ctx, input }) => {
-      return ctx.prisma.note.delete({
+      const note = await ctx.prisma.note.findFirst({
+        where: { id: input.id, userId: ctx.session.user.id },
+      });
+
+      if (!note) throw new TRPCError({ code: "NOT_FOUND" });
+
+      return await ctx.prisma.note.delete({
         where: { id: input.id },
-        include: {
+        select: {
+          id: true,
+          createdAt: true,
+          lastUpdated: true,
+          title: true,
+          userId: true,
           tags: { orderBy: { createdAt: "asc" } },
         },
       });
@@ -56,10 +118,21 @@ export const notesRouter = createTRPCRouter({
   rename: protectedProcedure
     .input(z.object({ id: z.string().cuid(), newTitle: z.string() }))
     .mutation(async ({ ctx, input }) => {
+      const note = await ctx.prisma.note.findFirst({
+        where: { id: input.id, userId: ctx.session.user.id },
+      });
+
+      if (!note) throw new TRPCError({ code: "NOT_FOUND" });
+
       return ctx.prisma.note.update({
         where: { id: input.id },
         data: { title: input.newTitle },
-        include: {
+        select: {
+          id: true,
+          createdAt: true,
+          lastUpdated: true,
+          title: true,
+          userId: true,
           tags: { orderBy: { createdAt: "asc" } },
         },
       });
@@ -68,12 +141,23 @@ export const notesRouter = createTRPCRouter({
   addTag: protectedProcedure
     .input(z.object({ id: z.string().cuid(), tagId: z.string().cuid() }))
     .mutation(async ({ ctx, input }) => {
+      const note = await ctx.prisma.note.findFirst({
+        where: { id: input.id, userId: ctx.session.user.id },
+      });
+
+      if (!note) throw new TRPCError({ code: "NOT_FOUND" });
+
       return ctx.prisma.note.update({
         where: { id: input.id },
         data: {
           tags: { connect: { id: input.tagId } },
         },
-        include: {
+        select: {
+          id: true,
+          createdAt: true,
+          lastUpdated: true,
+          title: true,
+          userId: true,
           tags: { orderBy: { createdAt: "asc" } },
         },
       });
@@ -82,16 +166,54 @@ export const notesRouter = createTRPCRouter({
   removeTag: protectedProcedure
     .input(z.object({ id: z.string().cuid(), tagId: z.string().cuid() }))
     .mutation(async ({ ctx, input }) => {
-      console.log("Got remove tag request");
+      const note = await ctx.prisma.note.findFirst({
+        where: { id: input.id, userId: ctx.session.user.id },
+      });
+
+      if (!note) throw new TRPCError({ code: "NOT_FOUND" });
 
       return ctx.prisma.note.update({
         where: { id: input.id },
         data: {
           tags: { disconnect: { id: input.tagId } },
         },
-        include: {
+        select: {
+          id: true,
+          createdAt: true,
+          lastUpdated: true,
+          title: true,
+          userId: true,
           tags: { orderBy: { createdAt: "asc" } },
         },
       });
+    }),
+
+  updateContent: protectedProcedure
+    .input(
+      z.object({
+        id: z.string().cuid(),
+        patches: patchesSchema,
+      })
+    )
+    .mutation(async ({ ctx, input }) => {
+      console.log("Got patches:", input.patches);
+
+      const note = await getFullNote(ctx.prisma, ctx.session, input.id);
+      if (!note) throw new TRPCError({ code: "NOT_FOUND" });
+
+      // compute and apply text patches
+      const dmp = new DMP.diff_match_patch();
+      const [newContent, results] = dmp.patch_apply(
+        input.patches,
+        note.content
+      );
+
+      // update note with new text content
+      const updatedNote = await ctx.prisma.note.update({
+        where: { id: input.id },
+        data: { content: newContent, lastUpdated: new Date() },
+        include: { tags: true },
+      });
+      return;
     }),
 });
